refactor: S3 클래스 패키지 이동
- 다른 도메인 패키지들과 마찬가지로, 레이어별로 패키지를 나눈다.
