+++ src/main/java/com/example/solidconnection/community/post/service/PostCommandService.java
@@ -11,11 +11,11 @@ import com.example.solidconnection.community.post.dto.PostUpdateRequest;
 import com.example.solidconnection.community.post.dto.PostUpdateResponse;
 import com.example.solidconnection.community.post.repository.PostRepository;
 import com.example.solidconnection.common.exception.CustomException;
-import com.example.solidconnection.s3.S3Service;
-import com.example.solidconnection.s3.UploadedFileUrlResponse;
+import com.example.solidconnection.s3.service.S3Service;
+import com.example.solidconnection.s3.dto.UploadedFileUrlResponse;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.repository.SiteUserRepository;
-import com.example.solidconnection.s3.ImgType;
+import com.example.solidconnection.s3.domain.ImgType;
 import com.example.solidconnection.community.post.domain.PostCategory;
 import com.example.solidconnection.util.RedisUtils;
 import lombok.RequiredArgsConstructor;
