+++ src/main/java/com/example/solidconnection/s3/dto/UploadedFileUrlResponse.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.s3;
+package com.example.solidconnection.s3.dto;
 
 public record UploadedFileUrlResponse(
         String fileUrl) {
