+++ src/main/java/com/example/solidconnection/siteuser/service/MyPageService.java
@@ -1,13 +1,13 @@
 package com.example.solidconnection.siteuser.service;
 
 import com.example.solidconnection.common.exception.CustomException;
-import com.example.solidconnection.s3.S3Service;
-import com.example.solidconnection.s3.UploadedFileUrlResponse;
+import com.example.solidconnection.s3.service.S3Service;
+import com.example.solidconnection.s3.dto.UploadedFileUrlResponse;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.dto.MyPageResponse;
 import com.example.solidconnection.siteuser.repository.LikedUniversityRepository;
 import com.example.solidconnection.siteuser.repository.SiteUserRepository;
-import com.example.solidconnection.s3.ImgType;
+import com.example.solidconnection.s3.domain.ImgType;
 import com.example.solidconnection.university.domain.LikedUniversity;
 import com.example.solidconnection.university.dto.UniversityInfoForApplyPreviewResponse;
 import lombok.RequiredArgsConstructor;
