+++ src/test/java/com/example/solidconnection/community/post/service/PostCommandServiceTest.java
@@ -12,9 +12,9 @@ import com.example.solidconnection.community.post.dto.PostUpdateResponse;
 import com.example.solidconnection.community.post.repository.PostImageRepository;
 import com.example.solidconnection.community.post.repository.PostRepository;
 import com.example.solidconnection.common.exception.CustomException;
-import com.example.solidconnection.s3.ImgType;
-import com.example.solidconnection.s3.S3Service;
-import com.example.solidconnection.s3.UploadedFileUrlResponse;
+import com.example.solidconnection.s3.domain.ImgType;
+import com.example.solidconnection.s3.service.S3Service;
+import com.example.solidconnection.s3.dto.UploadedFileUrlResponse;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.fixture.SiteUserFixture;
 import com.example.solidconnection.support.integration.BaseIntegrationTest;
